website/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
Django==4.1.3
django-versatileimagefield==2.2
djangorestframework==3.14.0
idna==3.4
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
Pillow==9.3.0
pycparser==2.21
python-decouple==3.6
python-magic==0.4.27
python-magic-bin==0.4.13
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.12
